+++ R/volcano4D.R
@@ -1,26 +1,79 @@
 #' Add mode bar button to rotate the plot
 #'
 #' @param p The volcano3D plot
-#' @param rotate_icon_path The icon path for rotation
-#' @param stop_icon_path The icon path for stop button
-#' @param rotate_colour The colour for the rotate button (default='lightgreen')
-#' @param stop_colour The colour for the stop button (default='tomato')
+#' @param rotate_icon_path The svg icon path for rotation. If NULL a play 
+#' button is used
+#' @param stop_icon_path The svg icon path for stop button. If NULL a pause 
+#' button is used
+#' @param rotate_colour The colour for the rotate button (default="#c7c7c7")
+#' @param stop_colour The colour for the stop button (default='#ff6347', 
+#' a.k.a 'tomato')
+#' @param scale Scaling for rotation button
+#' @param speed The rotation speed (default=180) 
+#' @param shiny_event_names If using shiny, pass in any shiny event names which
+#' should stop rotation when triggered (e.g. shiny_event_names = c('replot'))
 #' @importFrom htmlwidgets JS onRender
 #' @importFrom plotly config
+#' @return Returns a rotating cylindrical 3D plotly plot featuring variables on 
+#' a tri-axis radial graph with the -log10(multi-group test p-value) on the 
+#' z-axis
+#' @references
+#' Lewis, Myles J., et al. (2019).
+#' \href{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}{
+#' Molecular portraits of early rheumatoid arthritis identify clinical and 
+#' treatment response phenotypes.}
+#' \emph{Cell reports}, \strong{28}:9
+#' @importFrom plotly plot_ly add_trace add_text layout %>%
+#' @importFrom grDevices hsv rgb
+#' @keywords hplot iplot
+#' @concept volcanoplot 
+#' @export
+#' @examples
+#' data(example_data)
+#' syn_polar <- polar_coords(sampledata = syn_example_meta,
+#'                           contrast = "Pathotype", 
+#'                           groups = NULL, 
+#'                           pvalues = syn_example_p, 
+#'                           expression = syn_example_rld, 
+#'                           p_col_suffix = "pvalue", 
+#'                           padj_col_suffix = "padj", 
+#'                           non_sig_name = "Not Significant", 
+#'                           multi_group_prefix = "LRT",
+#'                           significance_cutoff = 0.01, 
+#'                           fc_col_suffix='log2FoldChange',
+#'                           fc_cutoff = 0.3)
+#' 
+#' p <- volcano3D(syn_polar, 
+#'     label_rows = c("FMOD", "LAMP5", "TNNT3"), 
+#'     xy_aspectratio = 1, 
+#'     label_size = 10, 
+#'     z_aspectratio = 0.9)
+#'     
+#' volcano4D(p)
 #' @export
 
 volcano4D <- function(p, 
                       rotate_icon_path=NULL, 
                       stop_icon_path=NULL, 
-                      rotate_colour='lightgreen', 
-                      stop_colour='tomato') {
+                      rotate_colour="#c7c7c7", 
+                      stop_colour='#ff6347', 
+                      scale='scale(0.4) translate(-4, -4)', 
+                      speed = 180, 
+                      shiny_event_names=c()) {
+  
+  if(class(p)[1] != "plotly") stop("p must be a 3D plotly plot")
   
   if(is.null(rotate_icon_path)){
-    rotate_icon_path <- "M19.432,7.157c-0.312-1.113-1.624-1.858-3.496-2.17c0.279,0.331,0.532,0.685,0.754,1.06c1.043,0.299,1.748,0.764,1.911,1.344c0.095,0.335,0.014,0.729-0.24,1.169c-0.274,0.476-0.768,1.007-1.455,1.542c0-0.034,0.005-0.067,0.005-0.101c0-3.816-3.094-6.91-6.91-6.91c-3.816,0-6.91,3.094-6.91,6.91c0,1.169,0.293,2.268,0.805,3.232c-1.366-0.277-2.303-0.805-2.495-1.487c-0.094-0.336-0.013-0.729,0.241-1.169c0.138-0.239,0.35-0.496,0.595-0.756c0.011-0.449,0.055-0.89,0.138-1.317c-1.398,1.144-2.112,2.386-1.805,3.476c0.338,1.205,1.845,1.98,3.968,2.24C5.8,15.854,7.774,16.91,10,16.91c3.389,0,6.201-2.44,6.791-5.659C18.735,9.951,19.795,8.448,19.432,7.157 M10,16.047c-1.651,0-3.147-0.664-4.238-1.738c0.147,0.005,0.295,0.008,0.447,0.008c1.502,0,3.195-0.212,4.941-0.658c1.734-0.443,3.297-1.064,4.595-1.776C14.952,14.299,12.682,16.047,10,16.047 M15.998,10.733c-1.27,0.797-2.973,1.554-5.062,2.088c-1.616,0.414-3.251,0.632-4.727,0.632c-0.427,0-0.827-0.025-1.213-0.061C4.338,12.425,3.954,11.258,3.954,10c0-3.339,2.707-6.046,6.046-6.046c3.34,0,6.047,2.708,6.047,6.046C16.047,10.249,16.027,10.492,15.998,10.733"
-  }
+    rotate_icon_path <- paste0("M20 33l12-9-12-9v18zm4-29C12.95 4 4 12.95 4", 
+                               " 24s8.95 20 20 20 20-8.95 20-20S35.05 4 24 ", 
+                               "4zm0 36c-8.82 0-16-7.18-16-16S15.18 8 24", 
+                               " 8s16 7.18 16 16-7.18 16-16 16z")
+    }
   
   if(is.null(stop_icon_path)){
-    stop_icon_path <- "M512 512H0V0h512v512z"
+    stop_icon_path <- paste0("M24 4C12.95 4 4 12.95 4 24s8.95 20 20 20 ", 
+                             "20-8.95 20-20S35.05 4 24 4zm-2 28h-4V16h4v16zm8", 
+                             " 0h-4V16h4v16z")
   }
   
   button_script <- readLines(system.file("/spin_function/spin_button.js",
@@ -28,22 +81,34 @@ volcano4D <- function(p,
     gsub("rotate_icon_path", rotate_icon_path, .)  %>% 
     gsub("stop_icon_path", stop_icon_path, .)  %>% 
     gsub("rotate_colour", rotate_colour, .)  %>% 
-    gsub("stop_colour", stop_colour, .)  
+    gsub("stop_colour", stop_colour, .) 
   
   rotate_button <- list(
     name = "rotate",
     title = 'Rotate',
     icon= list(
       path = rotate_icon_path,
-      transform = 'scale(0.84) translate(-1, 0)'
+      transform = scale
     ),
     val=FALSE, 
     attr='rotating',
     click = htmlwidgets::JS(button_script))
   
+  js_rotation <- readLines(system.file("spin_function/rotation.js", 
+                        package = "volcano3D")) %>%
+    gsub('rotation_speed', speed, .) %>%
+    gsub("rotate_colour", rotate_colour, .) 
+  
+  if(length(shiny_event_names) > 0){
+    print("catching")
+    event_name_str <- paste0("'", paste(shiny_event_names, collapse="', '"), 
+                             "'")
+    print(event_name_str)
+    js_rotation <- gsub("shiny_event_names", event_name_str, js_rotation)
+  }
+  
   p %>% 
     config(modeBarButtonsToAdd = list(rotate_button)) %>%
-    htmlwidgets::onRender(readLines(system.file("spin_function/rotation.js", 
-                                                package = "volcano3D")))
+    htmlwidgets::onRender(js_rotation)
 }