+++ inst/spin_function/rotation.js
@@ -2,9 +2,9 @@
 */
 function plot_rotate(el) {
   // Update the colour of the rotate button
-  $("[data-title='Rotate'] svg path").css("fill", "#7ac143");
+  $("[data-title='Rotate'] svg path").css("fill", "rotate_colour");
   
-   // get the last button added (-2 because of plotly button and 0 index)
+   // get the index for the rotate button
   let buttons = document.getElementsByClassName("modebar-btn");
   let titles = Array();
   for (let i = 0; i < buttons.length; i++) {
@@ -12,22 +12,31 @@ function plot_rotate(el) {
   }
   var useIndex = titles.indexOf('Rotate');
   
-  //Rotate the plot  
   function run() {
-    var modeBarButtons = document.getElementsByClassName("modebar-btn")[useIndex];
-    var shouldRotate = (modeBarButtons.getAttribute("data-val") == 'true');
-
-    if (shouldRotate) {
-      rotate("scene", Math.PI / 180);
-    }
-    requestAnimationFrame(run);  // Recursive call
+    rotate("scene", Math.PI / rotation_speed);
+    spinID = requestAnimationFrame(run);  // Recursive call
   }
 
-  run(); // Start the animation
+  // Start/stop the animation
+  $("[data-title='Rotate']").on("click", function() {
+  	var modeBarButtons = document.getElementsByClassName("modebar-btn")[useIndex];
+    var shouldRotate = (modeBarButtons.getAttribute("data-val") == 'true');
+    if (shouldRotate) {
+  	  spinID = requestAnimationFrame(run);
+  	} else {
+      cancelAnimationFrame(spinID);
+    }
+  });
+  
+  // stop animation if change in plot type
+  $(document).on('shiny:inputchanged', function(event) {
+    if([shiny_event_names].includes(event.name)) {
+      cancelAnimationFrame(spinID);
+    }
+  });
 
   function rotate(id, angle) {
     var eye0 = el.layout[id].camera.eye;
-    console.log(eye0);
     if(! eye0) {
       eye0 = {
         x: 1,
@@ -35,7 +44,6 @@ function plot_rotate(el) {
         z: 1
       };
     }
-    console.log(eye0);
     
     var rtz = xyz2rtz(eye0);
     rtz.t += angle;